+++ R/cooccurrences.R
@@ -91,13 +91,23 @@ setMethod("as.data.frame", "cooccurrences_bundle", function(x){
 #' e <- subset(c, ll >= 10.83)
 #' format(e)
 #' 
-#' # using pipe operator may be convenient
+#' # using pipe operator with subset
 #' cooccurrences("REUTERS", query = "oil") %>%
 #'   subset(!is.na(ll)) %>%
 #'   subset(!word %in% tm::stopwords("en")) %>%
 #'   subset(count_coi >= 5) %>%
 #'   subset(ll >= 10.83) %>%
 #'   format()
+#'   
+#' # generate datatables htmlwidget with buttons for export (Excel & more)
+#' # (alternatively use openxlsx::write.xlsx())
+#' interactive_table <- cooccurrences("REUTERS", query = "oil") %>%
+#'   format() %>%
+#'   DT::datatable(
+#'     extensions = "Buttons",
+#'     options = list(dom = 'Btip', buttons = c("excel", "pdf", "csv"))
+#'   )
+#' if (interactive()) show(interactive_table)
 setGeneric("cooccurrences", function(.Object, ...) standardGeneric("cooccurrences") )
 
 #' @rdname cooccurrences